Fix exception on first opening of Pepyatka.
Exception "Assertion Failed: Expected hash or Mixin instance, got [object Undefined] at ember.js:910" happens because App.ShowSpinnerWhileRendering is undefined. There was no explicit dependency on the JS file where this mixin is defined, so there was no guarantee that the JS file was loaded in the browser. The fix adds this dependency.