Correctly define Timeline.posts property.
This fixes a TODO in the code concerning a duplicate definition of Timeline.posts. The definition which was a parameter of Object.extend() defined a class property, not an instance property. The correct way to define an instance property is to override the init() function and place the definition there.